fix some bugs
If the text field is empty, now the program prints 'Empty field'.
If the text has less than 5 lemmas, now the program prints 'The text must have at least 5 words'.
If the text is written in the language, different from English, now the program prints 'The text must be in English'.